refactor: move can_history_trigger_func to models module
